Interpret refactored.
