+++ qiskit_nature/problems/second_quantization/molecular/molecular_problem.py
@@ -22,11 +22,15 @@ from qiskit.opflow import PauliSumOp
 from qiskit_nature.drivers import FermionicDriver, QMolecule
 from qiskit_nature.operators.second_quantization import SecondQuantizedOp
 from qiskit_nature.operators.second_quantization.qubit_converter import QubitConverter
-from qiskit_nature.results import EigenstateResult, ElectronicStructureResult, DipoleTuple
+from qiskit_nature.results import EigenstateResult, ElectronicStructureResult
 from qiskit_nature.transformers import BaseTransformer
-from .aux_fermionic_ops_builder import _create_all_aux_operators
-from .fermionic_op_builder import build_fermionic_op
-from .hopping_ops_builder import build_hopping_operators
+from qiskit_nature.problems.second_quantization.molecular.builders.aux_fermionic_ops_builder \
+    import _create_all_aux_operators
+from qiskit_nature.problems.second_quantization.molecular.builders.fermionic_op_builder import \
+    build_fermionic_op
+from qiskit_nature.problems.second_quantization.molecular.builders.hopping_ops_builder import \
+    build_hopping_operators
+from .result_interpreter import _interpret
 from ..base_problem import BaseProblem
 
 
@@ -64,8 +68,8 @@ class MolecularProblem(BaseProblem):
     def hopping_ops(self, qubit_converter: QubitConverter,
                     excitations: Union[str, int, List[int],
                                        Callable[[int, Tuple[int, int]],
-                                                List[Tuple[Tuple[int, ...], Tuple[int, ...]]]]
-                                       ] = 'sd',
+                                                List[Tuple[
+                                                    Tuple[int, ...], Tuple[int, ...]]]]] = 'sd',
                     ) -> Tuple[Dict[str, PauliSumOp], Dict[str, List[bool]],
                                Dict[str, Tuple[Tuple[int, ...], Tuple[int, ...]]]]:
         """Generates the hopping operators and their commutativity information for the specified set
@@ -90,7 +94,6 @@ class MolecularProblem(BaseProblem):
         q_molecule = cast(QMolecule, self.molecule_data)
         return build_hopping_operators(q_molecule, qubit_converter, excitations)
 
-    # TODO refactor by decomposing and eliminate ifs
     def interpret(self, raw_result: Union[EigenstateResult, EigensolverResult,
                                           MinimumEigensolverResult]) -> ElectronicStructureResult:
         """Interprets an EigenstateResult in the context of this transformation.
@@ -101,96 +104,7 @@ class MolecularProblem(BaseProblem):
         Returns:
             An electronic structure result.
         """
-        eigenstate_result = None
-        if isinstance(raw_result, EigenstateResult):
-            eigenstate_result = raw_result
-        elif isinstance(raw_result, EigensolverResult):
-            eigenstate_result = EigenstateResult()
-            eigenstate_result.raw_result = raw_result
-            eigenstate_result.eigenenergies = raw_result.eigenvalues
-            eigenstate_result.eigenstates = raw_result.eigenstates
-            eigenstate_result.aux_operator_eigenvalues = raw_result.aux_operator_eigenvalues
-        elif isinstance(raw_result, MinimumEigensolverResult):
-            eigenstate_result = EigenstateResult()
-            eigenstate_result.raw_result = raw_result
-            eigenstate_result.eigenenergies = np.asarray([raw_result.eigenvalue])
-            eigenstate_result.eigenstates = [raw_result.eigenstate]
-            eigenstate_result.aux_operator_eigenvalues = [raw_result.aux_operator_eigenvalues]
-
-        q_molecule = cast(QMolecule, self._molecule_data)
-        q_molecule_transformed = cast(QMolecule, self._molecule_data_transformed)
-
-        result = ElectronicStructureResult()
-        result.combine(eigenstate_result)
-        result.computed_energies = np.asarray([e.real for e in eigenstate_result.eigenenergies])
-        result.hartree_fock_energy = q_molecule.hf_energy
-        result.nuclear_repulsion_energy = q_molecule.nuclear_repulsion_energy
-        if q_molecule.nuclear_dipole_moment is not None:
-            dipole_tuple = tuple(x for x in q_molecule.nuclear_dipole_moment)
-            result.nuclear_dipole_moment = cast(DipoleTuple, dipole_tuple)
-        result.ph_extracted_energy = q_molecule_transformed.energy_shift.get(
-            "ParticleHoleTransformer", 0)
-        result.frozen_extracted_energy = q_molecule_transformed.energy_shift.get(
-            "FreezeCoreTransformer", 0)
-        if result.aux_operator_eigenvalues is not None:
-            # the first three values are hardcoded to number of particles, angular momentum
-            # and magnetization in this order
-            result.num_particles = []
-            result.total_angular_momentum = []
-            result.magnetization = []
-            result.computed_dipole_moment = []
-            result.ph_extracted_dipole_moment = []
-            result.frozen_extracted_dipole_moment = []
-            if not isinstance(result.aux_operator_eigenvalues, list):
-                aux_operator_eigenvalues = [result.aux_operator_eigenvalues]
-            else:
-                aux_operator_eigenvalues = result.aux_operator_eigenvalues  # type: ignore
-            for aux_op_eigenvalues in aux_operator_eigenvalues:
-                if aux_op_eigenvalues is None:
-                    continue
-                if aux_op_eigenvalues[0] is not None:
-                    result.num_particles.append(
-                        aux_op_eigenvalues[0][0].real)  # type: ignore
-
-                if aux_op_eigenvalues[1] is not None:
-                    result.total_angular_momentum.append(
-                        aux_op_eigenvalues[1][0].real)  # type: ignore
-
-                if aux_op_eigenvalues[2] is not None:
-                    result.magnetization.append(aux_op_eigenvalues[2][0].real)  # type: ignore
-
-                # the next three are hardcoded to Dipole moments, if they are set
-                if len(
-                        aux_op_eigenvalues) >= 6 and q_molecule.has_dipole_integrals:
-                    # check if the names match
-                    # extract dipole moment in each axis
-                    dipole_moment = []
-                    for moment in aux_op_eigenvalues[3:6]:
-                        if moment is not None:
-                            dipole_moment += [moment[0].real]  # type: ignore
-                        else:
-                            dipole_moment += [None]
-
-                    result.reverse_dipole_sign = q_molecule.reverse_dipole_sign
-                    result.computed_dipole_moment.append(cast(DipoleTuple,
-                                                              tuple(dipole_moment)))
-                    result.ph_extracted_dipole_moment.append(
-                        (q_molecule_transformed.x_dip_energy_shift.get(
-                            "ParticleHoleTransformer", 0),
-                         q_molecule_transformed.y_dip_energy_shift.get(
-                             "ParticleHoleTransformer", 0),
-                         q_molecule_transformed.z_dip_energy_shift.get(
-                             "ParticleHoleTransformer", 0)))
-
-                    result.frozen_extracted_dipole_moment.append(
-                        (q_molecule_transformed.x_dip_energy_shift.get(
-                            "FreezeCoreTransformer", 0),
-                         q_molecule_transformed.y_dip_energy_shift.get(
-                             "FreezeCoreTransformer", 0),
-                         q_molecule_transformed.z_dip_energy_shift.get(
-                             "FreezeCoreTransformer", 0)))
-
-        return result
+        return _interpret(self.molecule_data, self.molecule_data_transformed, raw_result)
 
     def get_default_filter_criterion(self) -> Optional[Callable[[Union[List, np.ndarray], float,
                                                                  Optional[List[float]]], bool]]:
