+++ qiskit_nature/problems/second_quantization/vibrational/builders/vibrational_op_builder.py
@@ -18,9 +18,8 @@ from qiskit_nature.drivers import WatsonHamiltonian
 from qiskit_nature.drivers.bosonic_bases import BosonicBasis, HarmonicBasis
 
 from qiskit_nature.operators.second_quantization.vibrational_op import VibrationalOp
-from qiskit_nature.problems.second_quantization.vibrational.vibrational_label_builder import \
-    _create_labels
-
+from qiskit_nature.problems.second_quantization.vibrational.builders.vibrational_label_builder \
+    import _create_labels
 
 logger = logging.getLogger(__name__)
