+++ test/circuit/library/ansatzes/test_chc.py
@@ -27,8 +27,8 @@ from qiskit_nature.circuit.library.ansatzes.utils.vibration_excitation_generator
 from qiskit_nature.mappers.second_quantization.direct_mapper import DirectMapper
 from qiskit_nature.operators.second_quantization.qubit_converter import QubitConverter
 from qiskit_nature.operators.second_quantization.vibrational_op import VibrationalOp
-from qiskit_nature.problems.second_quantization.vibrational.vibrational_label_builder import \
-    _create_labels
+from qiskit_nature.problems.second_quantization.vibrational.builders.vibrational_label_builder \
+    import _create_labels
 
 
 class TestCHCVSCF(QiskitNatureTestCase):
