+++ test/mappers/second_quantization/test_direct_mapper.py
@@ -19,7 +19,7 @@ from test.mappers.second_quantization.resources.reference_direct_mapper import R
 
 from qiskit_nature.drivers import GaussianForcesDriver
 from qiskit_nature.mappers.second_quantization import DirectMapper
-from qiskit_nature.problems.second_quantization.vibrational.vibrational_op_builder import \
+from qiskit_nature.problems.second_quantization.vibrational.builders.vibrational_op_builder import \
         _build_vibrational_op
